Remapped AuraLight namespace to application/lib, since we need to be able to namespace things other than controllers.
Trying to set up routing.  What I have is working, as far as it goes.
All addresses except /testpage will reroute to index.php and show the
FrontPage.  /testpage will route to the TestPageController and display
it.  The .htaccess file is apparently needed for redirecting incoming
requests.  Routes are currently being stored in share/config/routes.

I set up a Common/ directory and started on an SQL class and Log class.
I have been unable to implement custom configuration for log files since
the bootstrapper was successfully implemented, so I'd like to mess with
that some more.